Rename SearchResult component folder to SearchResults
